Create log_reader.gd
Originally intended to periodically scan a log file on disk, check for changes and append the new content to a control which could display text. Didn't function properly because Godot holds an open handle to its current log output file until the program exits, so this script is never able to open its own handle to read the file. Could still be useful for watching other files for updates and reading the new content accordingly.